Fix uint64 underflow in origin_time/hstrat_rank for zero-gen surfaces
Cast dstream_rank and dstream_S to signed Int64 before computing
origin_time (in trie postprocessor) and hstrat_rank, preventing
unsigned integer underflow when dstream_rank < dstream_S. This occurs
for zero-generations-elapsed surfaces after the S-1 trunk threshold
change. Apply the same fix in both Polars and Pandas implementations
so the validation decorator comparison passes.

Also fix _apply_empty_output_schema to use UInt32 for id and Int64 for
hstrat_rank, matching the non-empty output schema. Update docstrings
to reflect the correct output types.

+++ hstrat/dataframe/_surface_build_tree.py
@@ -115,17 +115,17 @@ def surface_build_tree(
         alife standard format, with the following columns:
 
         Required schema:
-        - 'id' : pl.UInt64
+        - 'id' : pl.UInt32
             - Unique identifier for each taxon (RE alife standard format).
         - 'ancestor_id' : pl.UInt64
             - Unique identifier for ancestor taxon  (RE alife standard format).
-        - 'hstrat_rank' : pl.UInt64
+        - 'hstrat_rank' : pl.Int64
             - Num generations elapsed for ancestral differentia.
             - Corresponds to `dstream_Tbar` - `dstream_S` for inner nodes.
             - Corresponds to `dstream_T` - 1 - `dstream_S` for leaf nodes.
 
         Optional schema:
-        - 'origin_time' : pl.UInt64
+        - 'origin_time' : pl.Int64
             - Estimated origin time for phylogeny nodes, in generations elapsed
               since founding ancestor.
